Add moment locales to PrimeFaces locale and load them when needed
This means users will not have to do anything anymore to load the
moment locales. Loading the PrimeFaces locale automatically loads
the corresponding moment locale.

However, this means that the "locale" parameters of <p:timeline>
and <p:schedule> need to be deprecated, as PF would automatically
use the current locale. If we want to keep support for configuring
a different locale for PF and for these widgets, we could also
setup the build so that it generates separate files for each moment
locale.